actually fixed finish condition
So! turns out, if the finish condition is overlapping with the floor, this adds a collision to the counter so I have to increment everything by 1. So, at the moment the finish condition is floating in the air to prevent this, and I'm trying to figure out how to replace the collisionshape2d with a sprite so that our end condition can be tailored depending on the level!